feat: refactor notify.
