+++ public/r/split-button-preview-demo.json
@@ -9,7 +9,7 @@
   "files": [
     {
       "path": "registry/example/split-button-preview-demo.tsx",
-      "content": "\"use client\"\n\nimport { toast } from \"@/components/ui/notify-provider\"\nimport { SplitButton } from \"@/components/ui/split-button\"\n\nexport default function SplitButtonPreviewDemo() {\n  const options = [\n    {\n      label: \"Copy SVG\",\n      onClick: () => {\n        toast.success(\"Copied SVG\", {\n          description: \"SVG code copied to clipboard\",\n        })\n      },\n    },\n    {\n      label: \"Copy Data URL\",\n      onClick: () => {\n        toast.success(\"Copied Data URL\", {\n          description: \"Data URL copied to clipboard\",\n        })\n      },\n    },\n    {\n      label: \"Download SVG\",\n      onClick: () => {\n        toast.success(\"Downloading SVG\", {\n          description: \"Your file is being downloaded...\",\n        })\n      },\n    },\n  ]\n\n  return (\n    <SplitButton options={options} mainAction={() => {}} size=\"sm\">\n      Download PNG\n    </SplitButton>\n  )\n}\n",
+      "content": "\"use client\"\n\nimport { toast } from \"@/components/ui/notify\"\nimport { SplitButton } from \"@/components/ui/split-button\"\n\nexport default function SplitButtonPreviewDemo() {\n  const options = [\n    {\n      label: \"Copy SVG\",\n      onClick: () => {\n        toast.success(\"Copied SVG\", {\n          description: \"SVG code copied to clipboard\",\n        })\n      },\n    },\n    {\n      label: \"Copy Data URL\",\n      onClick: () => {\n        toast.success(\"Copied Data URL\", {\n          description: \"Data URL copied to clipboard\",\n        })\n      },\n    },\n    {\n      label: \"Download SVG\",\n      onClick: () => {\n        toast.success(\"Downloading SVG\", {\n          description: \"Your file is being downloaded...\",\n        })\n      },\n    },\n  ]\n\n  return (\n    <SplitButton options={options} mainAction={() => {}} size=\"sm\">\n      Download PNG\n    </SplitButton>\n  )\n}\n",
       "type": "registry:example",
       "target": "components/split-button-preview-demo.tsx"
     }
