+++ public/r/toast-close-button-demo.json
@@ -6,7 +6,7 @@
   "files": [
     {
       "path": "registry/example/toast-close-button-demo.tsx",
-      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify-provider\"\n\nexport default function ToastCloseButtonDemo() {\n  return (\n    <div className=\"flex flex-wrap gap-2 md:gap-4\">\n      <Button\n        className=\"w-full md:w-auto\"\n        onClick={() =>\n          toast.success(\"Closable toast\", {\n            closable: true,\n          })\n        }\n      >\n        With Close Button\n      </Button>\n\n      <Button\n        className=\"w-full md:w-auto\"\n        onClick={() =>\n          toast.info(\"Non-closable toast\", {\n            closable: false,\n          })\n        }\n      >\n        Without Close Button\n      </Button>\n    </div>\n  )\n}\n",
+      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify\"\n\nexport default function ToastCloseButtonDemo() {\n  return (\n    <div className=\"flex flex-wrap gap-2 md:gap-4\">\n      <Button\n        className=\"w-full md:w-auto\"\n        onClick={() =>\n          toast.success(\"Closable toast\", {\n            closable: true,\n          })\n        }\n      >\n        With Close Button\n      </Button>\n\n      <Button\n        className=\"w-full md:w-auto\"\n        onClick={() =>\n          toast.info(\"Non-closable toast\", {\n            closable: false,\n          })\n        }\n      >\n        Without Close Button\n      </Button>\n    </div>\n  )\n}\n",
       "type": "registry:example",
       "target": "components/toast-close-button-demo.tsx"
     }
