+++ public/r/toast-custom-config-demo.json
@@ -6,7 +6,7 @@
   "files": [
     {
       "path": "registry/example/toast-custom-config-demo.tsx",
-      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify-provider\"\n\nexport default function ToastCustomConfigDemo() {\n  return (\n    <div className=\"flex gap-4\">\n      <Button\n        onClick={() => {\n          toast.info(`Toast `, {\n            maxToast: 3,\n            description: \"Max 3 toasts will show\",\n          })\n        }}\n      >\n        Max 3 Toasts\n      </Button>\n    </div>\n  )\n}\n",
+      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify\"\n\nexport default function ToastCustomConfigDemo() {\n  return (\n    <div className=\"flex gap-4\">\n      <Button\n        onClick={() => {\n          toast.info(`Toast `, {\n            maxToast: 3,\n            description: \"Max 3 toasts will show\",\n          })\n        }}\n      >\n        Max 3 Toasts\n      </Button>\n    </div>\n  )\n}\n",
       "type": "registry:example",
       "target": "components/toast-custom-config-demo.tsx"
     }
