+++ public/r/toast-demo.json
@@ -6,7 +6,7 @@
   "files": [
     {
       "path": "registry/example/toast-demo.tsx",
-      "content": "\"use client\"\n\nimport { cn } from \"@/lib/utils\"\nimport { Button } from \"@/components/ui/button\"\n\nimport { statusStyles, toast } from \"@/components/ui/notify-provider\"\n\nexport default function ToastDemo() {\n  return (\n    <div className=\"flex w-full flex-wrap justify-center gap-3 p-4\">\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"info\"])}\n        onClick={() => toast.info(\"Information\")}\n      >\n        Info Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"success\"])}\n        onClick={() => toast.success(\"Success\")}\n      >\n        Success Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"warning\"])}\n        onClick={() => toast.warning(\"Warning\")}\n      >\n        Warning Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"error\"])}\n        onClick={() => toast.error(\"Error\")}\n      >\n        Error Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"default\"])}\n        onClick={() => toast.default(\"Default Toast\")}\n      >\n        Default Toast\n      </Button>\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"loading\"])}\n        onClick={() => toast.loading(\"loading\")}\n      >\n        Loading Toast 1\n      </Button>\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"loading\"])}\n        onClick={() =>\n          toast.loading(\"loading\", {\n            loaderVariant: \"loader-2\",\n          })\n        }\n      >\n        Loading Toast 2\n      </Button>\n    </div>\n  )\n}\n",
+      "content": "\"use client\"\n\nimport { cn } from \"@/lib/utils\"\nimport { Button } from \"@/components/ui/button\"\n\nimport { statusStyles, toast } from \"@/components/ui/notify\"\n\nexport default function ToastDemo() {\n  return (\n    <div className=\"flex w-full flex-wrap justify-center gap-3 p-4\">\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"info\"])}\n        onClick={() => toast.info(\"Information\")}\n      >\n        Info Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"success\"])}\n        onClick={() => toast.success(\"Success\")}\n      >\n        Success Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"warning\"])}\n        onClick={() => toast.warning(\"Warning\")}\n      >\n        Warning Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"error\"])}\n        onClick={() => toast.error(\"Error\")}\n      >\n        Error Toast\n      </Button>\n\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"default\"])}\n        onClick={() => toast.default(\"Default Toast\")}\n      >\n        Default Toast\n      </Button>\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"loading\"])}\n        onClick={() => toast.loading(\"loading\")}\n      >\n        Loading Toast 1\n      </Button>\n      <Button\n        variant=\"outline\"\n        className={cn(\"w-32\", statusStyles[\"loading\"])}\n        onClick={() =>\n          toast.loading(\"loading\", {\n            loaderVariant: \"loader-2\",\n          })\n        }\n      >\n        Loading Toast 2\n      </Button>\n    </div>\n  )\n}\n",
       "type": "registry:example",
       "target": "components/toast-demo.tsx"
     }
