+++ public/r/toast-with-description-demo.json
@@ -6,7 +6,7 @@
   "files": [
     {
       "path": "registry/example/toast-with-description-demo.tsx",
-      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify-provider\"\n\nexport default function ToastWithDescriptionDemo() {\n  return (\n    <Button\n      onClick={() =>\n        toast.info(\"Event has been created\", {\n          description:\n            \"We've created your event and sent invitations to your guests.\",\n          duration: 5000,\n          position: \"bottom-right\",\n        })\n      }\n    >\n      Show Toast with Description\n    </Button>\n  )\n}\n",
+      "content": "\"use client\"\n\nimport { Button } from \"@/components/ui/button\"\n\nimport { toast } from \"@/components/ui/notify\"\n\nexport default function ToastWithDescriptionDemo() {\n  return (\n    <Button\n      onClick={() =>\n        toast.info(\"Event has been created\", {\n          description:\n            \"We've created your event and sent invitations to your guests.\",\n          duration: 5000,\n          position: \"bottom-right\",\n        })\n      }\n    >\n      Show Toast with Description\n    </Button>\n  )\n}\n",
       "type": "registry:example",
       "target": "components/toast-with-description-demo.tsx"
     }
