+++ registry.json
@@ -222,19 +222,24 @@
       "registryDependencies": ["utils"],
       "files": [
         {
-          "path": "registry/reusables/ui/notify.tsx",
+          "path": "registry/reusables/ui/notify/notify.tsx",
           "type": "registry:ui",
-          "target": "components/ui/notify.tsx"
+          "target": "components/ui/notify/notify.tsx"
         },
         {
-          "path": "registry/reusables/ui/notify-provider.tsx",
+          "path": "registry/reusables/ui/notify/notify-provider.tsx",
           "type": "registry:ui",
-          "target": "components/ui/notify-provider.tsx"
+          "target": "components/ui/notify/notify-provider.tsx"
         },
         {
-          "path": "registry/reusables/ui/notify-utils.ts",
+          "path": "registry/reusables/ui/notify/notify-utils.ts",
           "type": "registry:ui",
-          "target": "components/ui/notify-utils.ts"
+          "target": "components/ui/notify/notify-utils.ts"
+        },
+        {
+          "path": "registry/reusables/ui/notify/index.ts",
+          "type": "registry:ui",
+          "target": "components/ui/notify/index.ts"
         }
       ]
     },
