+++ registry/example/direct-image-upload-demo.tsx
@@ -3,7 +3,7 @@
 import { useFileUpload } from "@/hooks/use-file-upload"
 
 import { SingleImageDropzone } from "../reusables/single-image-upload"
-import { toast } from "../reusables/ui/notify-provider"
+import { toast } from "../reusables/ui/notify"
 
 export default function DirectImageUploadDemo() {
   const autoUpload = useFileUpload({
