+++ registry/example/multiple-image-upload-form-demo.tsx
@@ -22,7 +22,7 @@ import {
   FileState,
   MultiImageDropzone,
 } from "../reusables/multiple-image-upload"
-import { toast } from "../reusables/ui/notify-provider"
+import { toast } from "../reusables/ui/notify"
 
 const formSchema = z.object({
   title: z.string().min(2).max(50),
