+++ registry/example/split-button-preview-demo.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { toast } from "../reusables/ui/notify-provider"
+import { toast } from "../reusables/ui/notify"
 import { SplitButton } from "../reusables/ui/split-button"
 
 export default function SplitButtonPreviewDemo() {
